pyproject.toml
[tool.poetry]
name = "django-cassandra-example"
version = "0.1.0"
description = ""
authors = ["Arvind-4 <lordarvindthegamer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2"
python-dotenv = "^0.20.0"
django-cassandra-engine = "^1.7.0"
gunicorn = "^20.1.0"
uvicorn = "^0.17.6"
cassandra-driver = "^3.25.0"
whitenoise = "^6.0.0"
cryptography = "^36.0.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
